BGDIINF_SB-1180: improved logging
Instead of using the flask app logger which has always the same name
`app`, we now use a logger per module with the module name. This make
debugging easier because each message is then mapped to a python module.

Also fixed the pylint configuration in order to use old logging
formating style with %.