BGDIINF_SB-1181, BGDIINF_SB-1180: Do not pollute stdout/stderr while running unittest
When running unittest the stdout/stderr are polluted by logging message
(error message in case of an test case that test error). This is
annoying because on the console we see error message even all the tests
passed.

To avoid this we clear the logging handler and set the logging capture.
This way logging message are only displayed in case when test are
failing. We also need to use a whole logging configuration instead of
reusing the flask config.