BGDIINF_SB-1463: Fixed gevent/greenlet dependency issue
greenlet is a dependency of gevent and was not in the requirement
file. Due to this a new version of greenlet has been installed which is
not compatible with gevent.

To solve this gevent has been updated and now we use pipenv instead of
pip. Pipenv has the ability to freeze all dependencies version in
Pipfile.lock to avoid such issues.

Also added docker image metadata about git status of the image.

Also added isort to the make format target and fixed import sorting.