BGDIINF_SB-2420: Fixed browser origin validation
The browser doesn't add the Origin header when the request is made from the
same origin. However it does always send the Sec-Fetch-Site header telling
if the request is from the same origin, cross origin or from user.
See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Sec-Fetch-Site.
Unfortunately this Sec-Fetch-Site header is not supported by Safari !
Moreover we cannot hack the web application to always set the Origin header as
most browsers don't allow it.

So we need to check 2 headers: Sec-Fetch-Site and Origin with a fallback to the
Referer for Safari.

Also to have the correct CORS header in case of same origin we need to add the
correct allowed origin in CORS header which is the same as the request. In case
of cross-site then we use the Origin header as allowed origin in CORS.