BGDIINF_SB-2420: Fixed domain in CORS in case of 403 - #patch
In case a request was not accepted due to non allowed origin, the
Access-Control-Allow-Origin response header was not set correctly, it was set
to the request origin header, which indeed was not allowed.

Now in case of a 403 response we always use the request URL scheme and domain.
`Access-Control-Allow-Origin` cannot contained multiple domain but a single one
or `*`. Therefore if the request is not allowed we use the request URL domain.